+++ openscenario/openscenario_interpreter/src/syntax/act.cpp
@@ -21,7 +21,7 @@ namespace openscenario_interpreter
 inline namespace syntax
 {
 Act::Act(const pugi::xml_node & node, Scope & scope)
-: Scope(scope.makeChildScope(readAttribute<String>("name", node, scope))),
+: Scope(readAttribute<String>("name", node, scope), scope),
   start_trigger(readElement<Trigger>("StartTrigger", node, local()))
 {
   callWithElements(node, "ManeuverGroup", 1, unbounded, [&](auto && node) {