+++ simulation/traffic_simulator/include/traffic_simulator/math/hermite_curve.hpp
@@ -57,10 +57,10 @@ public:
   double getLength(size_t num_points) const;
   double getLength() const { return length_; }
   boost::optional<double> getSValue(
-    const geometry_msgs::msg::Point & point, size_t initial_resolution = 30,
-    size_t max_iteration = 30, double tolerance = 0.03, bool autoscale = false) const;
+    const geometry_msgs::msg::Pose & pose, double threshold_distance = 3.0,
+    bool autoscale = false) const;
   double getSquaredDistanceIn2D(
-    const geometry_msgs::msg::Point & point, double s, bool autoscale = false) const;
+    geometry_msgs::msg::Point point, double s, bool autoscale = false) const;
   boost::optional<double> getCollisionPointIn2D(
     const geometry_msgs::msg::Point & point0, const geometry_msgs::msg::Point & point1,
     bool search_backward = false) const;
@@ -71,11 +71,6 @@ public:
 private:
   std::pair<double, double> get2DMinMaxCurvatureValue() const;
   double length_;
-  boost::optional<double> solveNewtonMethod(
-    const geometry_msgs::msg::Point & point, double initial_value, double tolerance,
-    size_t max_iteration) const;
-  double getNewtonMethodStepSize(
-    const geometry_msgs::msg::Point & point, double s /*, bool autoscale = false*/) const;
 };
 }  // namespace math
 }  // namespace traffic_simulator