+++ docs/ReleaseNotes.md
@@ -3,10 +3,17 @@
 ## Difference between the latest release and master
 - Enable run t4v2_scenario_test with colcon test (with -DWITH_INTEGRATION_TEST=ON) ([link](https://github.com/tier4/scenario_simulator_v2/pull/531))
 
+## Version 0.5.8
+- Remove newton methods in getSValue function. ([link](https://github.com/tier4/scenario_simulator_v2/pull/612))
+- Set withLaneChange parameter as false. ([link](https://github.com/tier4/scenario_simulator_v2/pull/618))
+- Change traffic light topic name to "/perception/traffic_light_recognition/traffic_light_states" ([link](https://github.com/tier4/scenario_simulator_v2/pull/621))
+- Remove hard coded parameters in behavior tree plugin and use acceleration and deceleration value in traffic_simulator_msgs/msg/DriverModel. ([link](https://github.com/tier4/scenario_simulator_v2/pull/624))
+- Enable build with foxy latest version of behavior_tree_cpp_v3. ([link](https://github.com/tier4/scenario_simulator_v2/pull/625))
+- Add API::getDriverModel function. ([link](https://github.com/tier4/scenario_simulator_v2/pull/626))
+- Add Random test runner. [Link](https://github.com/tier4/scenario_simulator_v2-docs/pull/619) (Contribution by [Robotec.ai](https://robotec.ai/)).
+
 ## Version 0.5.7
 - Fix problem in getNormalVector function. (Contribution by [Utaro-M](https://github.com/Utaro-M)).
-- Fix problem in getSValue function, this function allowed to return s values under 0 or over 1. ([link](https://github.com/tier4/scenario_simulator_v2-docs/pull/603))
-
 ## Version 0.5.6
 - Fix context panel to display conditions' status. (Contribution by [Utaro-M](https://github.com/Utaro-M)).
 - Add NPC Behavior Plugin and Behavior-Tree Plugin for Vehicle and Pedestrian. ([link](https://github.com/tier4/scenario_simulator_v2-docs/pull/566))